Add support for PreHash ML-DSA (SHAKE256 only)
This commit adds two functions crypto_sign_signature_pre_hash
and crypto_sign_verify_pre_hash implementing the pre-hashing mode of ML-DSA.

The message to signed is formatted as
0x01 || ctxlen (1 byte) || ctx || oid || H(m)
where H(m) is the pre-hash and oid is the object identifier of the used hash
algorithm.

The ACVP client is adjusted to support the pre-hashing test cases for SHAKE256.
The other hash functions will be added separately.

Resolves https://github.com/pq-code-package/mldsa-native/issues/39

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>