Example: Add basic_deterministic example
This commit adds an example exercising the MLD_CONFIG_NO_RANDOMIZED_API option,
i.e., how to build and use mldsa-native without providing a randombytes()
implementation but instead supplying all randomness externally.

We use a small trick to be able to produce the same testvectors as in the
basic example: We extract the randomness that is generated by our
deterministic RNG (notrandombytes.c) during the randomized tests and hardcode
the randomness in the example (main.c).

Resolves https://github.com/pq-code-package/mldsa-native/issues/598

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>